You can now pick which leader to give the second action to, and it tracks that
Sadly, I had to kick reef to do the update here, and I can't figure out why. Making changes to the leaders array doesn't seem to trigger a `reef:signal` event, in this code or in the console. But reef gives us a way to kick it, so I'm doing so.