Add Eris, a homegrown unit testing framework
I tried to see if Jest or Jasmine or even QUnit could run in-browser without NPM, but none of them seem able to do so. So I made Eris here to bring us a little chaos.

By default Eris does nothing, but if you add `?test` to the URL it'll put the JS code through the paces and output results to the console.