Add structure pick list
When you build a structure, we now use a defined list to offer options. Structures in a settlement can now be clicked on to reveal information about that structure.

The structures present at this point are all from the rules as written, and most of them refer to rules that no longer exist. I wanted to get the baseline in so that I could see the changes in future commits. I expect a number of these to be removed, and most of them to be changed pretty radically.