Make saving/restoring actually work, which involved a huge refactor
Previously, I made activities as web components and put them on the page. They were view components and also models. This was a mistake, becuase it was easy for those view components to have state that wasn't savable, and lots of state had to get smashed together or rediscovered each time we showed the component.

Instead, we now have a clean separation between model objects (which are saved) and component objects (which are in the DOM). `Activity` is now a model object, composed of `ActivityDecision` objects, which are shown on-screen using `ActivitySheet` and `ActivityDecisionPanel` objects. There's lots and lots of tests for the models, ensuring that they're simpler to use. I further broke the activity templates into their own files, and I want to test those too, but have yet to do so.

Doing the above made save and restore work great, but I had to totally invert the logic of the program in a lot of places, so then I had to rebuilt it step by step to be data-driven instead of view-driven: view components are only created because models are present, and only update when the models change.

Other stuff that happened as I did fixes:
- Once decisions were a thing, the UI updated to emphasize them and collapse as they resolve and focus moves away from them
- Added "nudges" to log when the user does stuff outside of activities (like adding or destroying a structure), we now log those in special "Nudge" activities. If the most recent activity is already a nudge or an event, we'll log there, but otherwise we'll make a new activity to add it to.
- Made it impossible to add activities when one is already active. Resolve it to move on!