Make saveValue mean something
Before, no decision actally overrode `saveValue`, so while it was called to render form elements, it wasn't ever actually used to *save* anything. Now it is, which meant that I had to figure out that it was working backward: we should push the picked value (an object) through `saveValue`, and then push the return value of that to `saveAs`: `resolution`/`settlementId` should be the ID (the output of the `saveValue`) and `resolutionValue`/`settlementIdValue` should be the object.

Also, broke the assumption that providing a `saveValue`/`displayValue`/`summaryValue` method would create a `saveValues`/`displayValues`/`summaryValues` dictionary for you. I'm not actually using those objects, and they required that I eager-load all the options, which makes restoring saves hard (since we hydrate the activities before we set the data, so looking up possible settlements or the like won't work).