Rename actions
Work the Land -> Build Up. This always worked on the "above ability" and the new name is a mnemonic for that, while also being a broader term for what can happen in this activity.
Celebrary Holiday -> Cool Down. Similarly, this mnemonic name contrasts with the above, and is a much broader term to boot.