Relax decision-value validation when loading from save
Some values are permissible when selected but will later be impermissible, like an NPC who starts with 1 activity per turn (and is eligble to be trained) then gets trained (and is no longer eligible).

So now we log the weirdness and move on while loading, but in a normal set we'll still error out.