Add structure upgrades
When building structures, you can now declare that a previous structure is the "foundation", and you will start your build with 100% of the foundation's cost already built-in to the new structure. When the build is complete, the foundation will be removed (along with the building site), and you will end up with only the new structure.