First real feats, benefiting first actors each turn
These both add a consumable each turn, which gives a bonus to a roll and then gets used up. One benefits settlements and the other benefits leaders.

Also added the assurance feats, but they have no mechanics; you just ignore the roll and pick an outcome.

Also also flipped a fixture so that settlements go first, which has been true for a while.